requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
et-xmlfile==1.1.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
numpy==1.25.1
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
